feat: extend CrowdSec ban controls to analytics, geo-logs and alert tables
- extract the banned badge + ban/unban dropdown into a shared
  IpBanControls component and reuse it in the access-logs table
- add badge and shield action to Analytics Top IPs, Geo Logs Top IPs
  and IP-scoped Security alert rows; mount live decision updates on
  the top-IP tables for badge parity with access logs
- debug-logs IP filter now uses the shared isValidIp (fixes compressed
  IPv6 filters like ::1)
- add /security to the breadcrumb route labels (showed "Page")
- add cursor-pointer to the shared Button base so actions read as
  clickable under Tailwind v4 preflight
