+++ resources/components/access-logs/access-logs-table.tsx
@@ -9,9 +9,7 @@ import {
   ArrowUp,
   ChevronsUpDown,
   Columns3,
-  Loader2,
   Search,
-  ShieldBan,
 } from "lucide-react"
 import {
   Table,
@@ -29,7 +27,6 @@ import {
   DropdownMenu,
   DropdownMenuCheckboxItem,
   DropdownMenuContent,
-  DropdownMenuItem,
   DropdownMenuLabel,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
@@ -40,12 +37,9 @@ import { FiltersDrawer, FilterSection } from "@/components/ui/filters-drawer"
 import {
   useAccessLogs,
   useAccessLogFacets,
-  useBannedIps,
-  useBanIp,
-  useUnbanIp,
-  useCrowdsecStatus,
   useCrowdsecLiveUpdates,
 } from "@/lib/queries"
+import { IpBanControls } from "@/components/crowdsec/ip-ban-controls"
 import { useDebouncedValue } from "@/hooks/use-debounced-value"
 import { useIsMobile } from "@/hooks/use-mobile"
 import {
@@ -56,53 +50,9 @@ import {
   type SortOrder,
 } from "@/lib/api"
 import { cn, isMobileViewport } from "@/lib/utils"
-import { BAN_DURATIONS, isValidIp } from "@/lib/crowdsec"
+import { isValidIp } from "@/lib/crowdsec"
 
 const PAGE_SIZES = [10, 20, 50, 100, 200, 500, 1000] as const
-
-/** Ban/unban dropdown on the IP cell; hidden unless machine credentials
- *  enable write access on the CrowdSec integration. */
-function IpBanAction({ ip, banned }: { ip: string; banned: boolean }) {
-  const { data: status } = useCrowdsecStatus()
-  const ban = useBanIp()
-  const unban = useUnbanIp()
-  const isPending = ban.isPending || unban.isPending
-  if (!status?.write_enabled) return null
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <Button
-          variant="ghost"
-          size="icon-xs"
-          className="ml-1 align-middle text-muted-foreground"
-          disabled={isPending}
-          title={banned ? "Unban this IP" : "Ban this IP"}
-        >
-          {isPending ? <Loader2 className="animate-spin" /> : <ShieldBan />}
-        </Button>
-      </DropdownMenuTrigger>
-      <DropdownMenuContent align="start">
-        {banned ? (
-          <DropdownMenuItem onClick={() => unban.mutate(ip)}>
-            Unban {ip}
-          </DropdownMenuItem>
-        ) : (
-          <>
-            <DropdownMenuLabel>Ban {ip}</DropdownMenuLabel>
-            {BAN_DURATIONS.map((d) => (
-              <DropdownMenuItem
-                key={d.value}
-                onClick={() => ban.mutate({ ip, duration: d.value })}
-              >
-                {d.label}
-              </DropdownMenuItem>
-            ))}
-          </>
-        )}
-      </DropdownMenuContent>
-    </DropdownMenu>
-  )
-}
 
 const HTTP_METHODS = [
   "GET", "POST", "PUT", "PATCH", "DELETE", "HEAD", "OPTIONS", "CONNECT", "TRACE",
@@ -294,7 +244,6 @@ export function AccessLogsTable() {
   // Facet values are fetched lazily, on first open of either dropdown.
   const [facetsEnabled, setFacetsEnabled] = useState(false)
   const { data: facets } = useAccessLogFacets({ enabled: facetsEnabled })
-  const { data: bannedIps } = useBannedIps()
   useCrowdsecLiveUpdates()
   const search = useDebouncedValue(searchInput, 300)
   const ip = useDebouncedValue(ipInput, 300)
@@ -565,23 +514,7 @@ export function AccessLogsTable() {
                         className={cn(c.align === "right" && "text-right")}
                       >
                         {c.render(row)}
-                        {c.key === "ipAddress" && (
-                          <>
-                            {bannedIps?.has(row.ipAddress) && (
-                              <Badge
-                                variant="destructive"
-                                className="ml-2 align-middle"
-                                title="Active CrowdSec ban decision for this IP"
-                              >
-                                Banned
-                              </Badge>
-                            )}
-                            <IpBanAction
-                              ip={row.ipAddress}
-                              banned={!!bannedIps?.has(row.ipAddress)}
-                            />
-                          </>
-                        )}
+                        {c.key === "ipAddress" && <IpBanControls ip={row.ipAddress} />}
                       </TableCell>
                     ))}
                   </TableRow>
