+++ resources/components/analytics/top-ips-table.tsx
@@ -9,12 +9,15 @@ import {
 } from "@/components/ui/table"
 import { Skeleton } from "@/components/ui/skeleton"
 import { formatBytes, formatNumber } from "@/lib/api"
-import { useTopIpStats } from "@/lib/queries"
+import { useCrowdsecLiveUpdates, useTopIpStats } from "@/lib/queries"
+import { IpBanControls } from "@/components/crowdsec/ip-ban-controls"
 import { TablePaginationFooter, usePagedRows } from "./table-pagination"
 
 export function TopIpsTable() {
   const { data, isLoading } = useTopIpStats({ limit: 25 })
   const { pageItems, ...pagination } = usePagedRows(data?.items)
+  // Keep banned badges in sync with external cscli/console decisions.
+  useCrowdsecLiveUpdates()
 
   return (
     <Card>
@@ -40,7 +43,10 @@ export function TopIpsTable() {
               <TableBody>
                 {pageItems.map((row) => (
                   <TableRow key={row.ip_address}>
-                    <TableCell className="font-mono text-xs">{row.ip_address}</TableCell>
+                    <TableCell className="font-mono text-xs">
+                      {row.ip_address}
+                      <IpBanControls ip={row.ip_address} />
+                    </TableCell>
                     <TableCell>{row.country_code ?? "-"}</TableCell>
                     <TableCell>{row.city ?? "-"}</TableCell>
                     <TableCell className="text-right tabular-nums">{formatNumber(row.hits)}</TableCell>
